Reland Remove chrome_elf_import_unittests and add delayloads_unittests
This change relands
chromium-review.googlesource.com/c/chromium/src/+/1595219 but with a fix
for the Win7 test failure. The only difference between this CL and the
original CL is the edits to delayloads_unittests.cc.

This change is a follow-up to
https://chromium-review.googlesource.com/c/chromium/src/+/1551709. It
removes the chrome_elf_imports_unittests executable and combines the
tests with the similar tests in the more generic delayloads_unittests
executable. It also adds delayloads_unittests to the pipeline for
execution.

Bug: 948829
Change-Id: I9713d51f6393c5e134607644672d616a5ea30993
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1597222
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Cliff Smolinsky <cliffsmo@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#657691}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56ac67bfa6b6a436dd0520749c863d53c0133081